# specs/001-deployable-mvp/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: AgentBus Broker API
  version: 1.0.0
  description: |
    RESTful API for autonomous agent registration, discovery, direct messaging, and pub/sub eventing.
    
    **Authentication**: All endpoints require Entra ID JWT bearer token with `AgentBus.Agent` or `AgentBus.Admin` role.
    
    **Rate Limits**: MVP has no enforced rate limits. Enterprise upgrade adds per-agent quotas via Azure API Management.
  contact:
    name: AgentBus Team
    email: [email protected]
  license:
    name: MIT
servers:
  - url: https://agentbus-broker.example.com
    description: Production
  - url: https://agentbus-broker-dev.example.com
    description: Development
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/agents/{agentId}:
    get:
      operationId: getAgent
      summary: Get agent by ID
      description: Retrieve specific agent registration details
      tags:
        - Registration
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

    delete:
      operationId: deregisterAgent
      summary: Deregister agent
      description: |
        Removes agent from registry and deletes inbox queue. Agent must deregister itself (JWT principal ID must match).
        Admin role can deregister any agent.
      tags:
        - Registration
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '204':
          description: Agent successfully deregistered
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Forbidden - agent can only deregister itself (unless Admin role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Entra ID JWT token with `AgentBus.Agent` or `AgentBus.Admin` role in `roles` claim.
        Example: `Authorization: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOi...`
  parameters:
    AgentId:
      name: agentId
      in: path
      required: true
      description: Agent unique identifier
      schema:
        type: string
        minLength: 3
        maxLength: 64
        pattern: '^[a-zA-Z0-9\-]+$'
        example: logistics-coordinator-v1-abc123
  schemas:
...
    Agent:
      type: object
      required:
        - id
        - name
        - version
        - status
        - capabilities
        - messageTypes
        - identity
        - endpoints
        - timestamps
      properties:
        id:
          type: string
          description: Unique agent identifier (auto-generated)
          example: logistics-coordinator-v1-abc123
        name:
          type: string
          example: Logistics Coordinator
        version:
          type: string
          example: 1.0.0
        status:
          type: string
          enum:
            - active
            - inactive
          description: Current operational status
        capabilities:
          type: array
          items:
            type: string
          example:
            - schedule-drone
            - assign-route
        messageTypes:
          type: object
          properties:
            accepts:
              type: array
              items:
                type: string
            emits:
              type: array
              items:
                type: string
        identity:
          type: object
          required:
            - managedIdentityId
            - principalId
            - tenantId
          properties:
            managedIdentityId:
              type: string
              description: Azure Managed Identity resource ID
              example: /subscriptions/{sub}/resourceGroups/{rg}/providers/Microsoft.ManagedIdentity/userAssignedIdentities/agent-logistics
            principalId:
              type: string
              format: uuid
              description: Managed Identity principal ID (from JWT sub claim)
            tenantId:
              type: string
              format: uuid
              description: Entra ID tenant ID
        endpoints:
          type: object
          required:
            - inboxQueueName
          properties:
            inboxQueueName:
              type: string
              description: Service Bus inbox queue name
              example: agent-logistics-coordinator-v1-abc123-inbox
            healthCheckUrl:
              type: string
              format: uri
              nullable: true
        metadata:
          type: object
          additionalProperties:
            type: string
          nullable: true
        timestamps:
          type: object
          required:
            - registeredAt
            - lastHeartbeat
          properties:
            registeredAt:
              type: string
              format: date-time
              description: Registration timestamp (ISO 8601 UTC)
            lastHeartbeat:
              type: string
              format: date-time
              description: Last heartbeat timestamp (ISO 8601 UTC)
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - traceId
      properties:
        error:
          type: string
          description: Error type
          example: BadRequest
        message:
          type: string
          description: Human-readable error message
          example: Invalid agent version format - must be SemVer 2.0
        traceId:
          type: string
          description: Distributed trace ID for correlation
          example: 00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01
        details:
          type: object
          nullable: true
          description: Additional error details (validation errors, etc.)
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Unauthorized
            message: Missing or invalid JWT token
            traceId: 00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: NotFound
            message: Agent 'unknown-agent-id' not found
            traceId: 00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01
...
    ServiceUnavailable:
      description: Service temporarily unavailable (dependency failure)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: ServiceUnavailable
            message: Cosmos DB temporarily unavailable - retry after 5 seconds
            traceId: 00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01
      headers:
        Retry-After:
          description: Suggested retry delay in seconds
          schema:
            type: integer
            example: 5